refactor: rename `ResourceSpawner` to `ResourceController` and update related references for naming consistency.
